requirements.txt
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
python-dotenv==1.0.1
